refactor code

+++ JsonComparer/Contact.cs
@@ -5,7 +5,7 @@ namespace JsonComparer
     public sealed class Contact
     {
         [JsonPropertyName("mobileNumber")]
-        public int MobileNumber { get; set; }
+        public long MobileNumber { get; set; }
 
         [JsonPropertyName("email")]
         public string Email { get; set; }
